finetune/requirements.txt
# Core
torch==2.3.0
torchvision
torchaudio
transformers==4.51.3
datasets==3.4.1
accelerate==0.34.1
peft==0.11.1
bitsandbytes==0.43.1
sentencepiece==0.2.0
protobuf==3.20.3
wandb
cut_cross_entropy
hf_transfer
msgspec
tyro
xformers # optional; if it still breaks, remove it entirely
boto3==1.34.112
flash-attn==2.6.3
trltrl<0.9.0



# Unsloth
# unsloth # pin a known-good version; or the latest that supports HF 4.43
# unsloth_zoo
# RL extra (only needed for rlhf stage)
google-generativeai==0.7.2

# AWS
# boto3==1.34.129
sagemaker==2.219.0

# Misc
python-dotenv==1.0.1

"unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git"

--no-deps "xformers<0.0.27"

